requirements.txt
playwright>=1.44.0
playwright-stealth>=1.0.6
apscheduler>=3.10.4
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.1
pyyaml>=6.0.1
python-dotenv>=1.0.1
pandas>=2.2.2
aiofiles>=23.2.1
